Add AbstractMenuWriter class to simplify adding new IMenuWriter implementations
